+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New dark theme style
 - Add a boundary information above the focus
 - Information worker data are now displayed in a table
+- Add the possibility to use compute settings which doesn't have all the fields
 
 ### 13.8.0